Refactor RequestSenderService
Extract Class Refactoring applied to the request sender service. A base request sender service is created that handles general requests that are not specific to any microservice.

I then created a request sender for each microservice that inherits from the general request sender that encapsulates the logic surrounding requests to said microservice.